% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMultiAssayExperiment.R
\name{buildMultiAssayExperiment}
\alias{buildMultiAssayExperiment}
\title{A unifying pipeline function for building MultiAssayExperiment objects}
\usage{
buildMultiAssayExperiment(
  TCGAcode,
  dataType = c("RNASeqGene", "RNASeq2Gene", "RNASeq2GeneNorm", "miRNASeqGene", "CNASNP",
    "CNVSNP", "CNASeq", "CNACGH", "Methylation", "mRNAArray", "miRNAArray", "RPPAArray",
    "Mutation", "GISTIC"),
  runDate = "20160128",
  analyzeDate = "20160128",
  version,
  serialDir = "data/raw",
  outDataDir = "data/bits",
  mapDir = "data/maps",
  upload = FALSE,
  uploadFolder,
  update = TRUE,
  force = FALSE,
  include = c("colData", "sampleMap", "metadata")
)
}
\arguments{
\item{dataType}{A character vector of TCGA assay types}

\item{runDate}{A single string indicating the Firehose run date, usually
'20160128'}

\item{analyzeDate}{`character(1)` The GDAC Firehose analysis run date, only
'20160128' is supported}

\item{version}{`character(1)` A version string for versioning data runs (such
as "1.0.0")}

\item{serialDir}{The directory corresponding to the serialized data from
RTCGAToolbox}

\item{outDataDir}{`character(1)` The single string indicating piecewise data
product save location}

\item{mapDir}{A single string indicating the directory where sampleMaps
are stored from `saveMapData`}

\item{upload}{`logical(1)` (default FALSE) Whether to save data products to
the cloud infrastructure, namely the Azure Blob Storage location.}

\item{uploadFolder}{`character(1)` The folder where to upload the data
(default "staginghub")}

\item{update}{`logical(1)` (default TRUE) Whether to update the metadata data
from the data pooled by the function}

\item{force}{(logical) Whether to resave all sample maps to the map
directory}

\item{include}{`character()` A vector of metadata names to include. It must
include any or all of "colData", "sampleMap", or "metadata". This allows to
only publish changed data.}

\item{TCGAcodes}{`character(1)` A single TCGA cancer code}
}
\description{
This is the main function of the package. It brings together all of the
exported functions to load, create, and save data objects on ExperimentHub
and return documented metadata
}
