% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipeline.R
\name{buildMultiAssayExperiments}
\alias{buildMultiAssayExperiments}
\title{A unifying pipeline function for building MultiAssayExperiment objects}
\usage{
buildMultiAssayExperiments(TCGAcodes, dataType = c("RNASeqGene",
  "RNASeq2GeneNorm", "miRNASeqGene", "CNASNP", "CNVSNP", "CNASeq",
  "CNACGH", "Methylation", "mRNAArray", "miRNAArray", "RPPAArray",
  "Mutation", "GISTIC"), runDate = "20160128",
  analyzeDate = "20160128", serialDir = "data/raw",
  outDataDir = "data/bits", mapDir = "data/maps",
  metadataFile = "MAEOinfo.csv", upload = TRUE, force = FALSE)
}
\arguments{
\item{TCGAcodes}{A character vector of TCGA cancer codes}

\item{dataType}{A character vector of TCGA assay types}

\item{runDate}{A single string indicating the Firehose run date, usually
'20160128'}

\item{analyzeDate}{The GDAC Firehose analysis run date, only '20160128' is
supported}

\item{serialDir}{The directory corresponding to the serialized data from
RTCGAToolbox}

\item{outDataDir}{The single string indicating piecewise data product save
location}

\item{mapDir}{A single string indicating the directory where sampleMaps
are stored from `saveMapData`}

\item{metadataFile}{A single string pointing to the CSV file in which to
save metadata}

\item{upload}{logical (default TRUE) Whether to save data products to the
cloud infrastructure, namely AWS S3 ExperimentHub bucket}

\item{force}{(logical) Whether to resave all sample maps to the map
directory}
}
\description{
This is the main function of the package. It brings together all of the
exported functions to load, create, and save data objects on ExperimentHub
and return documented metadata
}
