% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curationAvailable}
\alias{curationAvailable}
\title{TCGA disease codes that have been curated}
\format{An object of class \code{character} of length 19.}
\source{
https://github.com/waldronlab/multiassayexperiment-tcga
}
\usage{
curationAvailable
}
\description{
A character vector indication what TCGA cancer codes have associated
curation data obtained from publications.
}
\examples{
data('curationAvailable')

}
\author{
Marcel Ramos \email{marcel.ramos@roswellpark.org}
}
\keyword{data}
