% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDirectories.R
\name{dataDirectories}
\alias{dataDirectories}
\title{Create necessary folders for pipeline}
\usage{
dataDirectories(create = FALSE)
}
\arguments{
\item{create}{logical (default FALSE) Whether to create folders in
current package directory}
}
\value{
A character vector of created directories
}
\description{
This convenience function automatically creates directories needed
for file processing in the pipeline.
}
