% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load and assemble MultiAssayExperiments from files}
\usage{
loadData(cancer, dataType, runDate, serialDir, mapDir, force)
}
\arguments{
\item{cancer}{A single character string indicating the TCGA disease code}

\item{dataType}{A character vector of TCGA assay types}

\item{runDate}{A single string indicating the Firehose run date, usually
'20160128'}

\item{serialDir}{The directory corresponding to the serialized data from
RTCGAToolbox}

\item{mapDir}{A single string indicating the directory where sampleMaps
are stored from `saveMapData`}

\item{force}{(logical) Whether to resave all sample maps to the map
directory}
}
\value{
A harmonized MultiAssayExperiment object
}
\description{
loadData takes cancer, data type, and storage inputs to piece together a
working MultiAssayExperiment object
}
