% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveNupload.R
\name{saveNupload}
\alias{saveNupload}
\title{A function to save serialized objects and upload to ExperimentHub}
\usage{
saveNupload(dataList, cancer, directory = "data/bits", upload = TRUE,
  fileExt = ".rda")
}
\arguments{
\item{dataList}{A List of experiment data for a MultiAssayExperiment}

\item{cancer}{A single string indicating the TCGA cancer code}

\item{directory}{The file location for saving serialized data pieces}
}
\value{
Function saves and uploads data to the ExperimentHub AWS S3 bucket
}
\description{
This function requires the user to obtain AWS CLI credentials to
ExperimentHub for it to work.
}
