% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRTCGAdata.R
\name{saveRTCGAdata}
\alias{saveRTCGAdata}
\title{A function to save data from RTCGAToolbox}
\usage{
saveRTCGAdata(runDate = "20160128", diseaseCode,
  dataType = c("RNASeqGene", "miRNASeqGene", "RNASeq2GeneNorm", "CNASNP",
  "CNVSNP", "CNASeq", "CNACGH", "Methylation", "Mutation", "mRNAArray",
  "miRNAArray", "RPPAArray", "GISTIC"), analyzeDate = "20160128",
  directory, force = FALSE)
}
\arguments{
\item{runDate}{The GDAC Firehose run date, only '20160128' is available}

\item{diseaseCode}{A single string indicating TCGA cancer code}

\item{dataType}{A character vector indicating the assay type(s)}

\item{analyzeDate}{The GDAC Firehose analyze run date, only '20160128' is
available}

\item{force}{logical (default FALSE) whether to force redownload of
resources}

\item{drectory}{A single string indicating the directory in which to save
intermediary datasets}
}
\description{
Saves intermediate data from the GDAC Firehose pipeline via RTCGAToolbox
}
