% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInfo.R
\name{updateInfo}
\alias{updateInfo}
\title{Update metadata from data bits}
\usage{
updateInfo(dataList, cancer, filePath = "MAEOinfo.csv")
}
\arguments{
\item{dataList}{A List of experiment data for a MultiAssayExperiment}

\item{cancer}{A single string indicating the TCGA cancer code}

\item{filePath}{A single string pointing to the file where metadata should
be saved}
}
\value{
Function saves a file in filePath
}
\description{
This functoin takes a list of data objects, cancer code, and a file path
to document the metadata in 'filePath'
}
