% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadAzure.R
\name{uploadAzure}
\alias{uploadAzure}
\title{Upload data to Azure Blob Store}
\usage{
uploadAzure(
  sas,
  url = .BIOCONDUCTOR_BLOB_STORE_URL,
  container,
  files,
  dataFolder,
  version,
  file_ext,
  package
)
}
\arguments{
\item{sas}{`character(1)` The SAS token used for authentication provided by
the Bioconductor Core Team}

\item{url}{`character(1)` The Azure Blob Store URL; by default, the value of
the internal `.BIOCONDUCTOR_BLOB_STORE_URL` constant is used.}

\item{container}{`character(1)` The name of the container on Azure}

\item{dataFolder}{`character(1)` The folder where all data versions are held}

\item{version}{`character(1)` The sub-folder corresponding to a particular
version of the data, e.g., "v2.1.0"}

\item{file_ext}{`character(1)` The file extension of files to be uploaded
without the leading dot, e.g., "rda"}

\item{package}{`character(1)` The name of the package for which data is
uploaded for, this will be the sub-folder within the Azure Blob Store
container.}
}
\description{
The function provides a programmatic interface to data upload using the
[`AzureStor`](https://cran.r-project.org/package=AzureStor) package.
}
