% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curationAvailable}
\alias{curationAvailable}
\title{TCGA disease codes that have been curated}
\format{
An named `character` vector of length 19
}
\source{
https://github.com/waldronlab/MultiAssayExperiment.TCGA
}
\usage{
data("curationAvailable")
}
\description{
A character vector indication what TCGA cancer codes have
  associated curation data obtained from publications.
}
\examples{

data("curationAvailable")

}
\author{
Marcel Ramos <marcel.ramos@roswellpark.org>
}
\keyword{data}
