% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInfo.R
\name{updateInfo}
\alias{updateInfo}
\title{Update metadata from data bits}
\usage{
updateInfo(dataList, cancer, folderPath, version)
}
\arguments{
\item{dataList}{A List of experiment data for a MultiAssayExperiment}

\item{cancer}{A single string indicating the TCGA cancer code}

\item{folderPath}{A single string pointing to the folder where metadata
information for each cancer is to be saved}

\item{version}{character(1) A version string for versioning data runs
(such as "1.0.0")}
}
\value{
A CSV file saved in the data folder
}
\description{
This function takes a list of data objects, cancer code, file path, and
version to document the metadata in 'filePath'
}
